Separate the crate into a lib.rs and a consuming shotgun.rs
